add ppermission logic

+++ app/Http/Controllers/CategoryController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Crypt;
 
@@ -41,21 +42,23 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        try {
-            $validated = $request->validate([
-                'name' => 'required|unique:categories,name',
-            ]);
+        if (Auth::user()->hasPermissionTo('make_category')) {
+            try {
+                $validated = $request->validate([
+                    'name' => 'required|unique:categories,name',
+                ]);
 
-            Category::create([
-                'name' => $validated['name'],
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ]);
-        } catch (Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+                Category::create([
+                    'name' => $validated['name'],
+                    'created_at' => Carbon::now(),
+                    'updated_at' => Carbon::now(),
+                ]);
+                return response()->json(['message' => 'Category created successfully'], 200);
+
+            } catch (Exception $e) {
+                return response()->json(['error' => $e->getMessage()], 500);
+            }
         }
-
-        return response()->json(['message' => 'Category created successfully'], 200);
     }
 
     /**
@@ -87,18 +90,20 @@ class CategoryController extends Controller
      */
     public function destroy(Request $request)
     {
-        try {
-            $categoryId = Crypt::decryptString($request->query('category'));
-        } catch (\Exception $e) {
-            abort(400, 'Invalid category token');
-        }
+        if (Auth::user()->hasPermissionTo('delete_category')) {
+            try {
+                $categoryId = Crypt::decryptString($request->query('category'));
+            } catch (\Exception $e) {
+                abort(400, 'Invalid category token');
+            }
 
-        try {
-            Category::find($categoryId)->delete();
-        } catch (Exception $e) {
-            abort(400, 'Category cannot be deleted');
-        }
+            try {
+                Category::find($categoryId)->delete();
+            } catch (Exception $e) {
+                abort(400, 'Category cannot be deleted');
+            }
 
-        return response()->json(['message' => 'Category deleted successfully'], 200);
+            return response()->json(['message' => 'Category deleted successfully'], 200);
+        }
     }
 }
