+++ app/Http/Controllers/ReviewController.php
@@ -165,24 +165,28 @@ class ReviewController extends Controller
     public function store(Request $request)
     {
         try {
-            if ($request->filled('review_id')) {$request['review_id'] = Crypt::decryptString($request->review_id);}
-            if ($request->filled('user_id')) {$request['user_id'] = Crypt::decryptString($request->user_id);}
+            if (Auth::user()->hasPermissionTo('send_review')) {
+                if ($request->filled('review_id')) {$request['review_id'] = Crypt::decryptString($request->review_id);}
+                if ($request->filled('user_id')) {$request['user_id'] = Crypt::decryptString($request->user_id);}
 
-            $validated = $request->validate([
-                'review_id' => [Rule::exists(Review::class, 'id'), 'required'],
-                'user_id' => [Rule::exists(User::class, 'id'), 'required'],
-                'comment.review_admin' => 'nullable|max:65535',
-            ]);
-
-            $review = Review::find($validated['review_id']);
-
-            if ($review->categorizedReview) {
-                $review->categorizedReview()->update([
-                    'user_id' => $validated['user_id'],
-                    'review_admin_comment' => $validated['comment']['review_admin'] ?? null,
+                $validated = $request->validate([
+                    'review_id' => [Rule::exists(Review::class, 'id'), 'required'],
+                    'user_id' => [Rule::exists(User::class, 'id'), 'required'],
+                    'comment.review_admin' => 'nullable|max:65535',
                 ]);
+
+                $review = Review::find($validated['review_id']);
+
+                if ($review->categorizedReview) {
+                    $review->categorizedReview()->update([
+                        'user_id' => $validated['user_id'],
+                        'review_admin_comment' => $validated['comment']['review_admin'] ?? null,
+                    ]);
+                } else {
+                    return response()->json(['error' => 'This review has not been categorized yet'], 500);
+                }
             } else {
-                return response()->json(['error' => 'This review has not been categorized yet'], 500);
+                abort(403, 'You do not have permission to send a review');
             }
         } catch (Exception $e) {
             return response()->json(['error' => $e->getMessage()], 500);
@@ -202,7 +206,15 @@ class ReviewController extends Controller
             abort(400, 'Invalid review token');
         }
 
-        $review = Review::with('categorizedReview')->find($reviewId);
+        if (Auth::user()->hasRole('department_admin')) {
+            $review = Review::with('categorizedReview')
+                ->whereHas('categorizedReview', function ($query) {
+                    $query->where('user_id', Auth::user()->id);
+                })
+                ->find($reviewId);
+        } else {
+            $review = Review::with('categorizedReview')->find($reviewId);
+        }
 
         return response()->json($review);
     }
@@ -237,17 +249,25 @@ class ReviewController extends Controller
                 'comment.department_admin' => 'nullable|max:65535',
             ]);
 
-            if (!empty($validated['category_id'])) {$categorizedReview->update(['category_id' => $validated['category_id']]);}
-            if (!empty($validated['status']['review'])) {$categorizedReview->update(['review_status' => $validated['status']['review']]);}
-            if (!empty($validated['status']['action'])) {$categorizedReview->update(['action_status' => $validated['status']['action']]);}
-            if (!empty($validated['status']['answer'])) {$categorizedReview->update(['answer_status' => $validated['status']['answer']]);}
-            if (!empty($validated['comment']['review_admin'])) {$categorizedReview->update(['review_admin_comment' => $validated['comment']['review_admin']]);}
-            if (!empty($validated['comment']['department_admin'])) {$categorizedReview->update(['department_admin_comment' => $validated['comment']['department_admin']]);}
+            if (Auth::user()->hasPermissionTo('categorizing_review')) {
+                if (!empty($validated['category_id'])) {$categorizedReview->update(['category_id' => $validated['category_id']]);}
+            }
+            if (Auth::user()->hasPermissionTo('flagging_review')) {
+                if (!empty($validated['status']['review'])) {$categorizedReview->update(['review_status' => $validated['status']['review']]);}
+                if (!empty($validated['status']['action'])) {$categorizedReview->update(['action_status' => $validated['status']['action']]);}
+                if (!empty($validated['status']['answer'])) {$categorizedReview->update(['answer_status' => $validated['status']['answer']]);}
+            }
+            if (Auth::user()->hasPermissionTo('comment_review')) {
+                if (!empty($validated['comment']['review_admin'])) {$categorizedReview->update(['review_admin_comment' => $validated['comment']['review_admin']]);}
+                if (!empty($validated['comment']['department_admin'])) {$categorizedReview->update(['department_admin_comment' => $validated['comment']['department_admin']]);}
+            }
+
+            return response()->json(['message' => 'Review updated successfully'], 200);
+
         } catch (Exception $e) {
             return response()->json(['error' => $e->getMessage()], 500);
         }
 
-        return response()->json(['message' => 'Review updated successfully'], 200);
     }
 
     /**
@@ -255,18 +275,22 @@ class ReviewController extends Controller
      */
     public function destroy(Request $request)
     {
-        try {
-            $reviewId = Crypt::decryptString($request->query('review'));
-        } catch (\Exception $e) {
-            abort(400, 'Invalid review token');
-        }
+        if (Auth::user()->hasPermissionTo('delete_review')) {
+            try {
+                $reviewId = Crypt::decryptString($request->query('review'));
+            } catch (\Exception $e) {
+                abort(400, 'Invalid review token');
+            }
 
-        try {
-            Review::find($reviewId)->delete();
-        } catch (Exception $e) {
-            abort(400, 'Review cannot be deleted');
-        }
+            try {
+                Review::find($reviewId)->delete();
+            } catch (Exception $e) {
+                abort(400, 'Review cannot be deleted');
+            }
 
-        return response()->json(['message' => 'Review deleted successfully'], 200);
+            return response()->json(['message' => 'Review deleted successfully'], 200);
+        } else {
+            abort(403, 'You do not have permission to delete a review');
+        }
     }
 }
